Convert helperQuestion file to typescript
